pps-ctrl: collect current samples from supply
Adds ability to read current samples from the supply. The firmware on
the PPS3205t-3s sucks and there is no way to perform back-to-back reads.
Instead the supply must be programmed from the host PC (a 24-bit write)
and then it sends back a similarly formated 24-bit packet containing the
current samples.

This patch refactors the code quite a bit to add this functionality,
including a new option '-l' which causes pps-ctrl to loop forever,
constantly writing the specified configuration to the supply and then
reading back the sampled data and printing it to STDOUT.

Breaking the loop with SIGINT (ctrl-c) gracefully exits after closing
the serial connection.

Signed-off-by: Mike Turquette <mturquette@linaro.org>